% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_binary_datatype.R
\name{convert_binary_datatype}
\alias{convert_binary_datatype}
\title{Convert Binary Data Type}
\usage{
convert_binary_datatype(datatype = c("NIFTI_TYPE_UINT8", "NIFTI_TYPE_INT32",
  "NIFTI_TYPE_FLOAT32"))
}
\arguments{
\item{datatype}{data type from GIFTI image}
}
\value{
List of length 2: with elements of size and what
}
\description{
Converts a data type to the size and \code{what} for
\code{\link{readBin}}, necessary for \code{Base64Binary} and
\code{GZipBase64Binary} formats
}

