% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_endian.R
\name{convert_endian}
\alias{convert_endian}
\title{Convert Endian from GIFTI}
\usage{
convert_endian(endian)
}
\arguments{
\item{endian}{character passed from GIFTI XML}
}
\value{
Character string
}
\description{
Converts Endian format from GIFTI
}

