% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_intent.R
\name{convert_intent}
\alias{convert_intent}
\title{Convert Intent}
\usage{
convert_intent(intent)
}
\arguments{
\item{intent}{(character) string of intent type}
}
\value{
A character string
}
\description{
Converts the \code{intent} field from a GIFTI image
to a more standard naming
}

