% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_matrix.R
\name{create_data_matrix}
\alias{create_data_matrix}
\title{Create Data Matrix}
\usage{
create_data_matrix(data, dims, ordering = c("RowMajorOrder",
  "ColumnMajorOrder"))
}
\arguments{
\item{data}{Data output from \code{\link{data_decoder}}}

\item{dims}{Dimensions of output}

\item{ordering}{Ordering of the data}
}
\value{
Matrix of Values
}
\description{
Creats
}

