% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_array_attributes.R
\name{data_array_attributes}
\alias{data_array_attributes}
\title{Data Array Attributes}
\usage{
data_array_attributes(darray)
}
\arguments{
\item{darray}{List of \code{xml_nodes} from GIFTI data array}
}
\value{
\code{data.frame} of attributes
}
\description{
Parses a list of XML data to get the attributes
}

