% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gifti_data.R
\name{download_gifti_data}
\alias{download_gifti_data}
\title{Download GIFTI Test Data}
\usage{
download_gifti_data(outdir = system.file(package = "gifti"),
  overwrite = FALSE, ...)
}
\arguments{
\item{outdir}{Output directory for test file directory}

\item{overwrite}{Should files be overwritten if already exist?}

\item{...}{additional arguments to \code{\link{download.file}}}
}
\value{
Vector of file names
}
\description{
Downloads GIFTI test data from
\url{https://www.nitrc.org/frs/download.php/411/BV_GIFTI_1.3.tar.gz}
}

