% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gifti_data.R
\name{have_gifti_test_data}
\alias{have_gifti_test_data}
\title{Check Presence of GIFTI Test Data}
\usage{
have_gifti_test_data(outdir = system.file(package = "gifti"))
}
\arguments{
\item{outdir}{Output directory for test file directory}
}
\value{
Logical indicator
}
\description{
Checks if GIFTI test data is downloaded
}

