% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_gifti.R
\name{is.gifti}
\alias{is.gifti}
\alias{is_gifti}
\title{Test if GIFTI}
\usage{
is.gifti(x)

is_gifti(x)
}
\arguments{
\item{x}{object to test}
}
\value{
Logical if \code{x} is GIFTI
}
\description{
Simple wrapper to determine if class is GIFTI
}
