% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_triangles.R
\name{surf_triangles}
\alias{surf_triangles}
\title{Make Triangles from GIfTI Image}
\usage{
surf_triangles(file)
}
\arguments{
\item{file}{File name of GIfTI image, usuall \code{surf.gii}}
}
\value{
List of values corresponding to the \code{data} element from
\code{\link{readgii}}
}
\description{
Creates Triangles for plotting in RGL from a GIfTI image
}

