% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gifti.R
\name{write_gifti}
\alias{write_gifti}
\title{Write .gii xml from "gifti" object}
\usage{
write_gifti(gii, out_file)
}
\arguments{
\item{gii}{The "gifti" object}

\item{out_file}{Where to write the new GIFTI file}
}
\description{
Writes a "gifti" object to a GIFTI file (ends in *.gii).
}
