% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gifti_from_template.R
\name{write_gifti_from_template}
\alias{write_gifti_from_template}
\title{Write GIFTI from Template}
\usage{
write_gifti_from_template(template_file, out_file, new_data = NULL)
}
\arguments{
\item{template_file}{Path to the template GIFTI file}

\item{out_file}{Where to write the new GIFTI file}

\item{new_data}{List of new data values. The length of \code{new_data}
should match the number of "DataArray"s in the template GIFTI file. Default
\code{NULL} will leave the data unchanged.}
}
\description{
Writes a GIFTI file identical to an existing file, except with
 certain values replaced.
}
