% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_decoder.R
\name{data_decoder}
\alias{data_decoder}
\title{Array Data Decoder}
\usage{
data_decoder(values, encoding = c("ASCII", "Base64Binary",
  "GZipBase64Binary"), datatype = NULL, endian = c("little", "big"))
}
\arguments{
\item{values}{text from XML of GIFTI image}

\item{encoding}{encoding of GIFTI values}

\item{datatype}{Passed to \code{\link{convert_binary_datatype}}}

\item{endian}{Endian to pass in \code{\link{readBin}}}
}
\value{
Vector of values
}
\description{
Decodes values from a GIFTI image
}

