% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gifti.R
\name{readgii}
\alias{readgii}
\title{Read GIFTI File}
\usage{
readgii(file)
}
\arguments{
\item{file}{Name of file to read}
}
\value{
List of values
}
\description{
Reads a GIFTI File and parses the output
}

